requirements.txt
absl-py==0.6.1
aiohttp==1.0.5
alabaster==0.7.10
amqp==2.4.0
anaconda-client==1.6.9
anaconda-navigator==1.7.0
anaconda-project==0.8.2
aniso8601==6.0.0
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.1
astropy==2.0.3
async-timeout==3.0.1
attrs==17.4.0
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
base58==1.0.3
beautifulsoup4==4.6.0
billiard==3.5.0.5
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.2
blinker==1.4
bokeh==0.12.13
boto==2.48.0
Bottleneck==1.2.1
celery==4.2.1
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
community==1.0.0b1
conda==4.4.10
conda-build==3.4.1
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.1.4
cycler==0.10.0
Cython==0.27.3
cytoolz==0.9.0
dask==0.16.1
datashape==0.5.4
decorator==4.2.1
discord.py==0.16.12
distributed==1.20.2
django-crispy-forms==1.7.2
docutils==0.14
ecdsa==0.13
entrypoints==0.2.3
et-xmlfile==1.0.1
-e git+https://github.com/mobolic/facebook-sdk.git@65ff582e77f7ed68b6e9643a7490e5dee2a1031b#egg=facebook_sdk
fastcache==1.0.2
feedparser==5.2.1
filelock==2.0.13
firebase==3.0.1
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Mail==0.9.1
flask-marshmallow==0.9.0
flask-mongoengine==0.9.5
Flask-REST-JSONAPI==0.27.0
Flask-RESTful==0.3.7
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gast==0.2.1.post0
gcloud==0.18.3
gevent==1.2.2
glob2==0.6
gmpy2==2.0.8
googleapis-common-protos==1.5.9
greenlet==0.4.12
grpcio==1.17.1
h5py==2.7.1
heapdict==1.0.0
hexdump==3.3
html5lib==1.0.1
httpie==1.0.2
httplib2==0.12.1
idna==2.6
imageio==2.2.0
imagesize==0.7.1
imutils==0.5.2
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.31.5
jupyterlab-launcher==0.10.2
jwcrypto==0.6.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kombu==4.2.2.post1
lazy-object-proxy==1.3.1
llvmlite==0.21.0
locket==0.2.0
lxml==4.1.1
Markdown==3.0.1
MarkupSafe==1.0
marshmallow==2.18.0
marshmallow-jsonapi==0.21.0
matplotlib==2.1.2
mccabe==0.6.1
mistune==0.8.3
mock==2.0.0
mongoengine==0.17.0
mpmath==1.0.0
msgpack-python==0.5.1
multidict==4.5.2
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
netaddr==0.7.19
networkx==2.1
nltk==3.2.5
nose==1.3.7
notebook==5.4.0
numba==0.36.2
numexpr==2.6.4
numpy==1.15.0
numpydoc==0.7.0
oauth2client==4.1.3
odo==0.5.1
olefile==0.45.1
opencv-python==3.4.2.17
openpyxl==2.4.10
packaging==16.8
pandas==0.23.3
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
patsy==0.5.0
pbr==5.1.3
pep8==1.7.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
pkginfo==1.4.1
pluggy==0.6.0
ply==3.10
pocketsphinx==0.1.15
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.3
ptyprocess==0.5.2
py==1.5.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
PyAudio==0.2.11
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pyflakes==1.6.0
pygame==1.9.4
Pygments==2.2.0
pylint==1.8.2
pymongo==3.8.0
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySDL2==0.9.6
PySocks==1.6.7
pytest==3.3.2
python-chess==0.24.2
python-dateutil==2.7.3
python-jwt==3.2.4
python-louvain==0.13
pytz==2018.5
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.3
qrcode==6.1
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
redis==3.2.1
requests==2.18.4
requests-toolbelt==0.9.1
rope==0.10.7
rq==1.0
rsa==4.0
ruamel-yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.4.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
SpeechRecognition==3.8.1
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
spyder==3.2.6
SQLAlchemy==1.2.1
sseclient==0.0.23
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
torch==1.0.0
torchvision==0.2.1
tornado==4.5.3
traitlets==4.3.2
typing==3.6.2
unicodecsv==0.14.1
urllib3==1.22
utils==0.9.0
uWSGI==2.0.18
vine==1.2.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==3.4
Werkzeug==0.14.1
widgetsnbextension==3.1.0
wrapt==1.10.11
WTForms==2.2.1
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.5
xlwt==1.2.0
zict==0.1.3
